+++ src/components/button-bar/button-bar.component.tsx
@@ -1,23 +1,14 @@
 import React from "react";
 import { SpaceProps } from "styled-system";
 import StyledButtonBar from "./button-bar.style";
+import ButtonBarContext, {
+  ButtonBarContextProps,
+} from "./__internal__/button-bar.context";
 
-export interface ButtonBarContextProps {
-  /** Apply fullWidth style to the button bar */
-  fullWidth?: boolean;
-  /** Defines an Icon position for buttons: "before" | "after" */
-  iconPosition?: "before" | "after";
-  /** Assigns a size to the buttons: "small" | "medium" | "large" */
-  size?: "small" | "medium" | "large";
-  /** Color variants for new business themes: "primary" | "secondary" | "tertiary" | "darkBackground" */
-  buttonType?: "primary" | "secondary" | "primary";
-}
 export interface ButtonBarProps extends ButtonBarContextProps, SpaceProps {
   /** Button or IconButton Elements, to be rendered inside the component */
   children: React.ReactNode;
 }
-
-export const ButtonBarContext = React.createContext<ButtonBarContextProps>({});
 
 export const ButtonBar = ({
   children,
